method PolyfaceQuery.boundaryOfVisibleSubset
