+++ common/changes/@bentley/geometry-core/EDLFacetSilhouette_2020-11-03-19-52.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/geometry-core",
+      "comment": "New method PolyfaceQuery.boundaryOfVisibleSubset",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/geometry-core",
+  "email": "69321059+EarlinLutz@users.noreply.github.com"
+}
