+++ core/geometry/src/test/polyface/PolyfaceQuery.test.ts
@@ -620,6 +620,41 @@ describe("ReOrientFacets", () => {
     expect(ck.getNumErrors()).equals(0);
   });
 
+  it("ComputeSilhouettes", () => {
+
+    const ck = new Checker();
+    const allGeometry: GeometryQuery[] = [];
+    let x0 = 0;
+    const solids = Sample.createClosedSolidSampler(true);
+    const defaultOptions = StrokeOptions.createForFacets();
+    // REMARK: (EDL Oct 2020) Mutter and grumble.  The builder does not observe shouldTriangulate !!!
+    for (const solid of solids) {
+      for (const viewVector of [Vector3d.create(0, 0, 1), Vector3d.create(1, -1, 1)]) {
+        const builder = PolyfaceBuilder.create(defaultOptions);
+        builder.addGeometryQuery(solid);
+        const mesh = builder.claimPolyface();
+        let y0 = 0;
+        if (ck.testType<IndexedPolyface>(mesh)) {
+          GeometryCoreTestIO.captureCloneGeometry(allGeometry, mesh, x0, y0, 0);
+          for (const selector of [0, 1, 2]) {
+            const edges = PolyfaceQuery.boundaryOfVisibleSubset(mesh, selector as (0 | 1 | 2), viewVector);
+            if (edges) {
+              y0 += 40.0;
+              GeometryCoreTestIO.captureCloneGeometry(allGeometry, edges, x0, y0, 0);
+              y0 += 20.0;
+              const chains = RegionOps.collectChains([edges]);
+              GeometryCoreTestIO.captureCloneGeometry(allGeometry, chains, x0, y0, 0);
+            }
+          }
+        }
+        x0 += 20;
+      }
+      x0 += 20;
+    }
+    GeometryCoreTestIO.saveGeometry(allGeometry, "Polyface", "ComputeSilhouettes");
+
+    expect(ck.getNumErrors()).equals(0);
+  });
 });
 
 type LoopOrParityLoops = Point3d[] | Point3d[][];
